# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TwoDays MVP API
  version: 0.1.0
  description: |
    情侣陪伴类 Web App MVP 接口文档。

    约定：
    - 所有业务接口统一前缀为 `/api/v1`。
    - 认证使用 Bearer JWT。
    - 金额使用整数存储，单位由币种决定；JPY 可直接使用日元，CNY 建议使用分。
    - 日期字段如 `checkinDate`、`messageDate`、`startedOn` 使用 `YYYY-MM-DD`。
    - 时间字段使用 ISO 8601 UTC 时间字符串。
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production placeholder
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 邮箱验证码登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "400":
          $ref: "#/components/responses/AuthCodeError"
        "409":
          $ref: "#/components/responses/ConflictError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    AuthCodeError:
      description: 验证码错误或过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConflictError:
      description: 业务冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    EnvelopeBase:
      type: object
      required: [data, error, requestId]
      properties:
        data:
          nullable: true
        error:
          nullable: true
          $ref: "#/components/schemas/ErrorObject"
        requestId:
          type: string
          example: req_123
...
    ErrorEnvelope:
      type: object
      required: [data, error, requestId]
      properties:
        data:
          nullable: true
          example: null
        error:
          $ref: "#/components/schemas/ErrorObject"
        requestId:
          type: string
    ErrorObject:
      type: object
      nullable: true
      required: [code, message, details]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - VALIDATION_ERROR
            - AUTH_CODE_EXPIRED
            - AUTH_CODE_INVALID
            - RATE_LIMITED
            - INVITE_CODE_INVALID
            - RELATIONSHIP_FULL
            - ALREADY_IN_RELATIONSHIP
            - CHECKIN_ALREADY_EXISTS
            - GOAL_NOT_FOUND
            - RELATIONSHIP_INACTIVE
            - NOT_FOUND
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    LoginRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
        inviteCode:
          type: string
          nullable: true
          example: A8X2K
...
    AuthEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              required: [accessToken, refreshToken, user]
              properties:
                accessToken:
                  type: string
                refreshToken:
                  type: string
                user:
                  $ref: "#/components/schemas/User"
                relationship:
                  nullable: true
                  $ref: "#/components/schemas/RelationshipSummary"
...
    User:
      type: object
      required: [id, email, nickname, avatarUrl, timezone, locale, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nickname:
          type: string
          example: ゆうた
        avatarUrl:
          type: string
          nullable: true
        timezone:
          type: string
          example: Asia/Shanghai
        locale:
          type: string
          example: zh-CN
        createdAt:
          type: string
          format: date-time
...
    RelationshipSummary:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RelationshipStatus"
...
    RelationshipStatus:
      type: string
      enum: [pending, active, broken]
...